requirements.txt
apache-airflow-providers-snowflake